+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added Sei Network
 - Make the broadcast function public on Sender
 - Added a `call_as` helper on TxHandler Trait
+- Added optional disable of sorting of ExecuteFns and QueryFns fields by @Kayanski
+
 
 ## v0.15.0